Move utils module to constraints subpackage
