+++ baybe/constraints/utils.py
